+++ database/migrations/2023_11_10_111922_add_attributes_in_books_table.php
@@ -9,19 +9,39 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('books', function (Blueprint $table) {
-            $table->unsignedDecimal('budget', 12)->nullable()->after('bank_account_id');
-            $table->string('report_periode_code', 20)->default(Book::REPORT_PERIODE_IN_MONTHS)->after('budget');
-            $table->string('start_week_day_code', 10)->default('monday')->after('report_periode_code');
-        });
+        if (!Schema::hasColumn('books', 'budget')) {
+            Schema::table('books', function (Blueprint $table) {
+                $table->unsignedDecimal('budget', 12)->nullable()->after('bank_account_id');
+            });
+        }
+        if (!Schema::hasColumn('books', 'report_periode_code')) {
+            Schema::table('books', function (Blueprint $table) {
+                $table->string('report_periode_code', 20)->default(Book::REPORT_PERIODE_IN_MONTHS)->after('budget');
+            });
+        }
+        if (!Schema::hasColumn('books', 'start_week_day_code')) {
+            Schema::table('books', function (Blueprint $table) {
+                $table->string('start_week_day_code', 10)->default('monday')->after('report_periode_code');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('books', function (Blueprint $table) {
-            $table->dropColumn('budget');
-            $table->dropColumn('report_periode_code');
-            $table->dropColumn('start_week_day_code');
-        });
+        if (Schema::hasColumn('books', 'budget')) {
+            Schema::table('books', function (Blueprint $table) {
+                $table->dropColumn('budget');
+            });
+        }
+        if (Schema::hasColumn('books', 'report_periode_code')) {
+            Schema::table('books', function (Blueprint $table) {
+                $table->dropColumn('report_periode_code');
+            });
+        }
+        if (Schema::hasColumn('books', 'start_week_day_code')) {
+            Schema::table('books', function (Blueprint $table) {
+                $table->dropColumn('start_week_day_code');
+            });
+        }
     }
 };
